FLEXLOOKUP
FLEXLOOKUP is a Flexible XLOOKUP LAMBDA function that performs sorting under the hood when the binary search is called (setting the search mode to -2 or +2).
When the search mode is not set to binary, FLEXLOOKUP does introduce overhead by unnecessarily sorting the data. 
Binary search is significantly more performant for big datasets, especially on less performant machines. 
The "Time Execution" button runs a macro that reruns each formula "n" times where "n" is defined in cell B6. Might take minutes to run based on the inputs.
The execution times are printed in the top row.